refactor: Modal 관련 타입 파일 분리
