+++ components/src/lib/Modal/ConfirmModal/ConfirmModal.tsx
@@ -2,7 +2,7 @@ import MainModal from '../MainModal/MainModal';
 import Title from '../../components/Title/Title';
 import Button from '../../components/Button/Button';
 import { ModalHeader, ModalBody, ModalFooter } from '../../components/ModalLayout/ModalLayout';
-import type { ModalProps } from '../../types/common';
+import type { ModalProps } from '../types/Modal.type';
 
 export interface ConfirmModalProps extends ModalProps {
   headerText: string;
