+++ components/src/lib/Modal/PromptModal/PromptModal.tsx
@@ -3,7 +3,7 @@ import Title from '../../components/Title/Title';
 import Input from '../../components/Input/Input';
 import Button from '../../components/Button/Button';
 import { ModalHeader, ModalBody, ModalFooter } from '../../components/ModalLayout/ModalLayout';
-import type { ModalProps } from '../../types/common';
+import type { ModalProps } from '../types/Modal.type';
 
 export interface PromptModalProps extends ModalProps {
   headerText: string;
