+++ components/src/lib/Modal/types/Modal.type.ts
@@ -1,0 +1,13 @@
+export type Size = 'sm' | 'md' | 'lg';
+export type Position = 'center' | 'bottom' | 'top';
+export type BackDropType = 'transparent' | 'blur' | 'opaque';
+
+export interface ModalProps {
+  isOpen: boolean;
+  close: () => void;
+  position?: Position;
+  backdropType?: BackDropType;
+  size?: Size;
+  shadow?: boolean;
+  animation?: boolean;
+}
