+++ components/src/lib/types/common.ts
@@ -1,17 +1,3 @@
-export type Size = 'sm' | 'md' | 'lg';
-export type Position = 'center' | 'bottom' | 'top';
-export type BackDropType = 'transparent' | 'blur' | 'opaque';
-
 export type StrictPropsWithChildren<P = unknown> = P & {
   children: React.ReactNode;
 };
-
-export interface ModalProps {
-  isOpen: boolean;
-  close: () => void;
-  position?: Position;
-  backdropType?: BackDropType;
-  size?: Size;
-  shadow?: boolean;
-  animation?: boolean;
-}
